pyproject.toml
[tool.poetry]
name = "tlp"
version = "0.1.1"
description = "utility api for text-as-data course"
authors = ["Thurston <thurston.b.sexton@gmail.com>"]
license = "MIT"
include = [".dvc/config"]

[tool.poetry.dependencies]
python = "^3.9"
toolz = "^0.11.2"
pandera = "^0.8.0"
pyjanitor = "^0.22.0"
dvc = "^2.9.3"
markovify = "^0.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
